Task 3 0x0C LLP

+++ 0x0C-more_malloc_free/3-array_range.c
@@ -10,8 +10,8 @@
   */
 int *array_range(int min, int max)
 {
-	int i;
-	int range;
+	long int i;
+	long int range;
 	int *ptr;
 
 	if (min > max)
